compile fixes and some stuff in thread.rs

+++ libjvm/src/util.rs
@@ -3,7 +3,7 @@ use nix::errno::Errno::EINTR;
 pub fn retry_on_eintr(to_retry: impl Fn() -> i32) -> i32 {
     loop {
         let err = to_retry();
-        if nix::errno::errno() != EINTR || err != -1 {
+        if nix::errno::errno() != EINTR as i32 || err != -1 {
             return err;
         }
     }
